% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donotdisturb.R
\name{donotdisturb}
\alias{donotdisturb}
\title{Centered Headline 0 for LaTeX}
\usage{
donotdisturb()
}
\arguments{
\item{willdonothing}{String for nothing}
}
\description{
Gives centered comments for LaTeX. Always capitalized.
}
\examples{
function(willdonothing = "makesnosense")

}
