% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meowscraper_tiny.R
\name{find_tinyCat}
\alias{find_tinyCat}
\title{Random cat find image URL (faster)}
\usage{
find_tinyCat()
}
\arguments{
\item{NA}{None yet}
}
\description{
Returns URL of a randomly selected cat image from Wikimedia.org
}
\examples{
find_tinyCat()

}
