% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translators.R
\name{humaner}
\alias{humaner}
\title{Centered Headline 0 for LaTeX}
\usage{
humaner()
}
\arguments{
\item{willdonothing}{String for nothing}
}
\description{
Gives centered comments for LaTeX. Always capitalized.
}
\examples{
humaner("Meeeowww... Meowmeow?! Meowow.")

}
