% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meow_world.R
\name{meow_world}
\alias{meow_world}
\title{Centered Headline 0 for LaTeX}
\usage{
meow_world()
}
\arguments{
\item{willdonothing}{String for nothing}
}
\description{
Gives centered comments for LaTeX. Always capitalized.
}
\examples{
meow_world(willdonothing = "makesnosense")

}
