% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meower.R
\name{meower}
\alias{meower}
\title{Centered Headline 0 for LaTeX}
\usage{
meower()
}
\arguments{
\item{willdonothing}{String for nothing}
}
\description{
Gives centered comments for LaTeX. Always capitalized.
}
\examples{
meower("My name is Catty McCatface")

}
