% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pawr-package.R
\docType{package}
\name{PawR-package}
\alias{PawR}
\alias{PawR-package}
\title{A package from cats for cats!}
\description{
An open-source (GPL-3) R package that lets you translate cat language,
find the awesome new cats, and much more!!!
}
\keyword{internal}
