% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meowscraper.R
\name{show_randomCat}
\alias{show_randomCat}
\title{Random cat image URL}
\usage{
show_randomCat()
}
\arguments{
\item{NA}{None yet}
}
\description{
Returns the URL of a randomly selected cat image from Wikimedia.org
}
\examples{
show_randomCat()

}
