% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meowscraper_tiny.R
\name{show_tinyCat}
\alias{show_tinyCat}
\title{Random cat image URL (faster)}
\usage{
show_tinyCat()
}
\arguments{
\item{NA}{None yet}
}
\description{
Returns the URL of a randomly selected cat image from Wikimedia.org
}
\examples{
show_tinyCat()

}
